Add support for SA Vilebloom
 - reorganize colorway.js because it has gotten too big
   - break it into imports
