+++ configurator/src/scss/pantone.scss
@@ -1,3 +1,7 @@
 $color-pantone-141C: rgb(242,199,92);
 $color-pantone-533C: rgb(31,42,68);
 $color-pantone-534C: rgb(27,54,93);
+$color-pantone-177C: rgb(255,128,139);
+$color-pantone-2046U: rgb(205,108,143);
+$color-pantone-474U: rgb(246, 189, 157);
+$color-pantone-7699C: rgb(52, 101, 127);
